initial commit of port of spree-product-assembly to Solidus; requires fix to Solidus frontend/app/controllers/spree/orders_controller.rb
