+++ Gemfile
@@ -1,8 +1,8 @@
 source "https://rubygems.org"
 
 branch = 'master'
-gem 'solidus', github: 'solidusio/solidus', branch: branch
-
+gem 'solidus', github: 'solidusio/solidus', branch: "v1.2"
+gem 'deface', '~> 1.0', '>= 1.0.2'
 gem 'pry-rails'
 
 gemspec
