+++ README.markdown
@@ -40,21 +40,6 @@ The store will treat assemblies a bit different than regular products on checkou
 Spree will create and track inventory units for its parts rather than for the product itself.
 That means you essentially have a product composed of other products. From a customer perspective it's like they are paying a single amount for a collection of products.
 
-## Using with spree_wombat
-
-If you use this with spree_wombat make sure that you add this extension after spree_wombat in your `Gemfile`
-
-This extension provides a specific serializer for shipments `assembly_shipment_serializer`, to use this in your Spree storefront make sure you configure spree_wombat like this:
-
-```ruby
-config.payload_builder = {
-  'Spree::Shipment' => {
-    serializer: 'Spree::Wombat::AssemblyShipmentSerializer',
-    root: 'shipments'
-  }
-}
-```
-
 ---
 
 ## Contributing
