+++ app/assets/javascripts/spree/backend/solidus_product_assembly/index.js.coffee
