+++ app/assets/javascripts/spree/backend/solidus_product_assembly/translations.js.erb
