+++ app/assets/javascripts/spree/frontend/solidus_product_assembly.js
