+++ app/assets/stylesheets/spree/backend/solidus_product_assembly.css
