+++ app/assets/stylesheets/spree/frontend/solidus_product_assembly.css
