+++ app/models/spree/stock/inventory_unit_builder_decorator.rb
@@ -4,26 +4,16 @@ module Spree
       def units
         @order.line_items.flat_map do |line_item|
           line_item.quantity_by_variant.flat_map do |variant, quantity|
-            quantity.times.map { build_inventory_unit(variant, line_item) }
+            quantity.times.map {
+              @order.inventory_units.build(
+              pending: true,
+              variant: variant,
+              line_item: line_item,
+              order: @order
+              )
+            }
           end
         end
-      end
-
-      def build_inventory_unit(variant, line_item)
-        @order.inventory_units.includes(
-          variant: {
-            product: {
-              shipping_category: {
-                shipping_methods: [:calculator, { zones: :zone_members }]
-              }
-            }
-          }
-        ).build(
-          pending: true,
-          variant: variant,
-          line_item: line_item,
-          order: @order
-        )
       end
     end
   end
