+++ app/views/spree/admin/orders/_assemblies.html.erb
@@ -4,7 +4,6 @@
       <%= Spree.t(:product_bundles) %>
     </legend>
   </fieldset>
-
   <table class="product-bundles" data-hook="">
     <colgroup>
       <col style="width: 10%;" />
@@ -14,7 +13,6 @@
       <col style="width: 15%;" />
       <col style="width: 15%;" />
     </colgroup>
-
     <thead>
       <th colspan="2"><%= Spree.t(:item_description) %></th>
       <th><%= Spree.t(:price) %></th>
@@ -22,41 +20,40 @@
       <th><%= Spree.t(:total) %></th>
       <th class="orders-actions actions" data-hook="admin_order_form_line_items_header_actions"></th>
     </thead>
-
-    <% order.line_items.assemblies.each do |item| %>
-      <tbody>
-        <tr class="" data-item-quantity="<%= item.quantity %>">
-          <td class="item-image"><%= mini_image(item.variant) %></td>
-          <td class="item-name">
-            <%= item.product.name %> - <%= item.variant.sku %> <br />
-            <%= "(" + variant_options(item.variant) + ")" unless item.variant.option_values.empty? %>
+    <tbody>
+      <% order.line_items.assemblies.each do |item| %>
+        <%= content_tag :tr, class: 'line-item', id: "line-item-#{item.id}", data: { line_item_id: item.id } do %>
+          <td class="line-item-image"><%= image_tag(item.variant.display_image.attachment(:mini)) %></td>
+          <td class="line-item-name">
+            <%= item.product.name %><br><%= "(" + variant_options(item.variant) + ")" unless item.variant.option_values.empty? %>
+            <% if item.variant.sku.present? %>
+              <strong><%= Spree.t(:sku) %>:</strong> <%= item.variant.sku %>
+            <% end %>
           </td>
-          <td class="item-price align-center">
+          <td class="line-item-price align-center">
             <%= item.single_money.to_html %>
           </td>
           <td class="line-item-qty-show align-center">
             <%= item.quantity %>
           </td>
-
-          <td class="line-item-qty-edit hidden">
-            <%= number_field_tag :quantity, item.quantity, :min => 0, class: "line_item_quantity", :size => 5 %>
+          <td class="line-item-qty-edit">
+            <%= number_field_tag :quantity, item.quantity, min: 0, class: "line_item_quantity", size: 5 %>
           </td>
-
-          <td class="item-total align-center">
+          <td class="line-item-total align-center">
             <%= item.display_amount.to_html %>
           </td>
-          <td class="cart-item-delete actions" data-hook="cart_item_delete">
+          <td class="cart-line-item-delete actions" data-hook="cart_item_delete">
             <% unless item.any_units_shipped? %>
               <% if can? :update, item %>
-              <%= link_to '', '#', :class => 'save-line-item fa fa-ok no-text with-tip', :data => { 'line-item-id' => item.id, :action => 'save'}, :title => Spree.t('actions.save'), :style => 'display: none' %>
-              <%= link_to '', '#', :class => 'cancel-line-item fa fa-cancel no-text with-tip', :data => {:action => 'cancel'}, :title => Spree.t('actions.cancel'), :style => 'display: none' %>
-              <%= link_to '', '#', :class => 'edit-line-item fa fa-edit no-text with-tip', :data => {:action => 'edit'}, :title => Spree.t('edit') %>
-              <%= link_to '', '#', :class => 'delete-line-item fa fa-trash no-text with-tip', :data => { 'line-item-id' => item.id, :action => 'remove'}, :title => Spree.t('delete') %>
+              <%= link_to '', '#', class: 'save-line-item fa fa-ok no-text with-tip', title: Spree.t('actions.save') %>
+              <%= link_to '', '#', class: 'cancel-line-item fa fa-cancel no-text with-tip', title: Spree.t('actions.cancel') %>
+              <%= link_to '', '#', class: 'edit-line-item fa fa-edit no-text with-tip', title: Spree.t('edit') %>
+              <%= link_to '', '#', class: 'delete-line-item fa fa-trash no-text with-tip', title: Spree.t('delete') %>
               <% end %>
             <% end %>
           </td>
-        </tr>
-      </tbody>
-    <% end %>
+        <% end %>
+      <% end %>
+    </tbody>
   </table>
 <% end %>
