+++ lib/solidus_product_assembly.rb
@@ -1,0 +1,4 @@
+require 'solidus_core'
+require 'solidus_product_assembly/engine'
+require 'coffee_script'
+require 'sass/rails'
