--- lib/spree_product_assembly.rb
@@ -1,4 +1,0 @@
-require 'spree_core'
-require 'spree_product_assembly/engine'
-require 'coffee_script'
-require 'sass/rails'
