+++ solidus_product_assembly.gemspec
@@ -1,7 +1,7 @@
 Gem::Specification.new do |s|
   s.platform    = Gem::Platform::RUBY
-  s.name        = 'soludus_product_assembly'
-  s.version     = '1.2.0.beta'
+  s.name        = 'solidus_product_assembly'
+  s.version     = '1.2.0.rc2'
   s.summary     = 'Adds opportunity to make bundle of products to your Solidus store'
   s.description = s.summary
   s.required_ruby_version = '>= 1.9.3'
@@ -15,19 +15,19 @@ Gem::Specification.new do |s|
   s.require_path = 'lib'
   s.requirements << 'none'
 
-  s.add_dependency 'solidus_backend', '~> 1.2.0'
+  s.add_dependency 'solidus_backend', s.version
 
-  s.add_development_dependency 'active_model_serializers', '0.9.0.alpha1'
-  s.add_development_dependency 'capybara', '~> 2.4'
-  s.add_development_dependency 'coffee-rails', '~> 4.0.0'
-  s.add_development_dependency 'database_cleaner', '~> 1.4'
-  s.add_development_dependency 'factory_girl', '~> 4.4'
+  # s.add_development_dependency 'active_model_serializers', '0.9.0.alpha1'
+  s.add_development_dependency 'capybara', '~> 2.6'
+  s.add_development_dependency 'coffee-rails', '~> 4.1'
+  s.add_development_dependency 'database_cleaner', '~> 1.5'
+  s.add_development_dependency 'factory_girl', '~> 4.5'
   s.add_development_dependency 'ffaker'
   s.add_development_dependency 'launchy'
   s.add_development_dependency 'pg'
-  s.add_development_dependency 'poltergeist', '~> 1.6'
-  s.add_development_dependency 'rspec-rails', '~> 3.3.0'
-  s.add_development_dependency 'sass-rails', '~> 4.0.0'
+  s.add_development_dependency 'poltergeist', '~> 1.8'
+  s.add_development_dependency 'rspec-rails', '~> 3.4'
+  s.add_development_dependency 'sass-rails', '~> 5.0'
   s.add_development_dependency 'simplecov'
   s.add_development_dependency 'sqlite3'
 end
